// lib/shopify/types.ts
export type Connection<T> = {
   edges: Array<Edge<T>>;
};
// ...
export type Image = {
   url: string;
   altText: string;
   width: number;
   height: number;
};

export type Menu = {
   title: string;
   path: string;
};

export type Money = {
   amount: string;
   currencyCode: string;
};

export type Page = {
   id: string;
   title: string;
   handle: string;
   body: string;
   bodySummary: string;
   seo?: SEO;
   createdAt: string;
   updatedAt: string;
};

export type Product = Omit<ShopifyProduct, 'variants' | 'images'> & {
   variants: ProductVariant[];
   images: Image[];
};

export type ProductOption = {
   id: string;
   name: string;
   values: string[];
};

export type ProductVariant = {
   id: string;
   title: string;
   availableForSale: boolean;
   selectedOptions: {
      name: string;
      value: string;
   }[];
   price: Money;
};

export type SEO = {
   title: string;
   description: string;
};

export type ShopifyCart = {
   id: string;
   checkoutUrl: string;
   cost: {
      subtotalAmount: Money;
      totalAmount: Money;
      totalTaxAmount: Money;
   };
   lines: Connection<CartItem>;
   totalQuantity: number;
};

export type ShopifyCollection = {
   handle: string;
   title: string;
   description: string;
   seo: SEO;
   updatedAt: string;
   image: Image;
};

export type ShopifyProduct = {
   id: string;
   handle: string;
   availableForSale: boolean;
   title: string;
   description: string;
   descriptionHtml: string;
   options: ProductOption[];
   compareAtPriceRange: {
      maxVariantPrice: Money;
      minVariantPrice: Money;
   };
   priceRange: {
      maxVariantPrice: Money;
      minVariantPrice: Money;
   };
   variants: Connection<ProductVariant>;
   featuredImage: Image;
   images: Connection<Image>;
   seo: SEO;
   tags: string[];
   updatedAt: string;
};